requirements.txt
numpy
pandas
matplotlib
nibabel
scipy
sckit-image
lmfit
antspyx
